Relocates App.tsx outside of components dir
