+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDom from "react-dom/client";
-import App from "./components/App";
+import App from "./App";
 
 const rootElement = document.getElementById("root");
